+++ src/test/java/MapReduceTest.java
@@ -17,9 +17,9 @@ import java.util.List;
 
 public class MapReduceTest {
 
-    private MapDriver<IntWritable, Text, Text, IntWritable> mapDriver;
+    private MapDriver<LongWritable, Text, Text, IntWritable> mapDriver;
     private ReduceDriver<Text, IntWritable, Text, IntWritable> reduceDriver;
-    private MapReduceDriver<IntWritable, Text, Text, IntWritable, Text, IntWritable> mapReduceDriver;
+    private MapReduceDriver<LongWritable, Text, Text, IntWritable, Text, IntWritable> mapReduceDriver;
 
     @Before
     public void setUp() {
@@ -46,7 +46,7 @@ public class MapReduceTest {
     @Test
     public void testMapper() throws IOException {
         mapDriver
-                .withInput(new IntWritable(), new Text("1, 76123, 10"))
+                .withInput(new LongWritable(), new Text("1, 76123, 10"))
                 .withOutput(new Text("metricName, 60000"), new IntWritable(10))
                 .runTest();
     }
@@ -66,10 +66,10 @@ public class MapReduceTest {
     @Test
     public void testMapReduce() throws IOException {
         mapReduceDriver
-                .withInput(new IntWritable(), new Text("1, 145947, 30"))
-                .withInput(new IntWritable(), new Text("1, 74123, 40"))
-                .withInput(new IntWritable(), new Text("2, 12345, 5"))
-                .withInput(new IntWritable(), new Text("2, 23456, 15"))
+                .withInput(new LongWritable(), new Text("1, 145947, 30"))
+                .withInput(new LongWritable(), new Text("1, 74123, 40"))
+                .withInput(new LongWritable(), new Text("2, 12345, 5"))
+                .withInput(new LongWritable(), new Text("2, 23456, 15"))
                 .withOutput(new Text("metricName, 120000, 1m"), new IntWritable(30))
                 .withOutput(new Text("metricName, 60000, 1m"), new IntWritable(40))
                 .withOutput(new Text("secondName, 0, 1m"), new IntWritable(10))
